Move Phrase indexer into index package
