+++ kifi-backend/modules/search/app/com/keepit/search/SearchGlobal.scala
@@ -9,7 +9,7 @@ import com.keepit.search.index.article.ArticleIndexerPlugin
 import play.api.Mode._
 import play.api._
 import com.keepit.search.index.user.UserIndexerPlugin
-import com.keepit.search.phrasedetector.PhraseIndexerPlugin
+import com.keepit.search.index.phrase.PhraseIndexerPlugin
 import com.keepit.search.index.graph.collection.CollectionGraphPlugin
 import com.keepit.search.index.graph.user._
 import com.keepit.search.index.graph.library.LibraryIndexerPlugin
