+++ kifi-backend/modules/search/app/com/keepit/search/controllers/search/ArticleIndexerController.scala
@@ -8,7 +8,7 @@ import com.keepit.search.IndexInfo
 import com.keepit.search.index.article.ArticleIndexer
 import com.keepit.search.index.article.ArticleIndexerPlugin
 import com.keepit.search.index.Indexer.CommitData
-import com.keepit.search.phrasedetector.PhraseIndexer
+import com.keepit.search.index.phrase.PhraseIndexer
 import com.keepit.shoebox.ShoeboxServiceClient
 import org.apache.lucene.document.Document
 import play.api.libs.json._
