+++ kifi-backend/modules/search/app/com/keepit/search/index/phrase/PhraseFieldBuilder.scala
@@ -1,4 +1,4 @@
-package com.keepit.search.phrasedetector
+package com.keepit.search.index.phrase
 
 import com.keepit.search.Lang
 import com.keepit.search.index.DefaultAnalyzer
